New folder name Utils is made and constants file is moved in that folder
